+++ Makefile
@@ -1,16 +1,6 @@
-# COLORS
-GREEN = \033[0;32m
-RED = \033[0;31m
-RESET = \033[0m
-
 CC = cc
 CFLAGS = -Wall -Werror -Wextra
 LIBFT_PATH = libft
-# Add any other static library in the same fashion - 1
-
-# TODO: reorganize the Makefile and the folder structure
-# INCLUDE: .h files or .a library folders
-# SRC: .c files
 SRC_DIR = srcs
 OBJ_DIR = obj
 INCLUDE_DIRS = include -I $(LIBFT_PATH) -I /opt/homebrew/Cellar/readline/8.2.1/include
@@ -23,11 +13,35 @@ INCLUDE_DIRS +=	-I /opt/vagrant/embedded/include
 LIBS	+= -L/opt/vagrant/embedded/lib
 endif
 
-SRC_FILES = $(wildcard $(SRC_DIR)/*.c)
+SRC_FILES = srcs/builtins_2.c \
+			srcs/builtins.c \
+			srcs/cd.c \
+			srcs/echo.c \
+			srcs/execute.c \
+			srcs/expand.c \
+			srcs/export.c \
+			srcs/ft_free.c \
+			srcs/heredoc.c \
+			srcs/init.c \
+			srcs/main.c \
+			srcs/pipe_redirect_separator_2.c \
+			srcs/pipe_redirect_separator.c \
+			srcs/pipe.c \
+			srcs/qvalidate.c \
+			srcs/redirection.c \
+			srcs/rvalidate_2.c \
+			srcs/rvalidate.c \
+			srcs/signals.c \
+			srcs/start_exit.c \
+			srcs/to_double_pointer.c \
+			srcs/unset.c \
+			srcs/utils_2.c \
+			srcs/utils.c \
+			srcs/where.c
+
 OBJ_FILES = $(patsubst $(SRC_DIR)/%.c,$(OBJ_DIR)/%.o,$(SRC_FILES))
 
 LIBFT = $(LIBFT_PATH)/libft.a
-# Add any other static library in the same fashion - 2
 
 NAME = minishell
 
@@ -37,28 +51,22 @@ all: $(NAME)
 
 $(NAME): $(OBJ_FILES) $(LIBFT)
 	@$(CC) $(CFLAGS) $(LIBS) -o $@ $^
-	@echo "$(GREEN)+ $(NAME)$(RESET)"
-# Add any other static library in the same fashion as $(LIBFT) AND -I$(LIBFT_PATH) - 3
-
+	@echo Making $(NAME)
 
 $(OBJ_DIR)/%.o: $(SRC_DIR)/%.c
 	@mkdir -p $(OBJ_DIR)
 	@$(CC) $(CFLAGS) -I $(INCLUDE_DIRS) -c -o $@ $<
-# Add any other static library in the same fashion as -I$(LIBFT_PATH) - 4
 
 $(LIBFT):
 	@$(MAKE) -C $(LIBFT_PATH)
-# Add any other static library in the same fashion - 5
 
 clean:
 	@$(MAKE) -C $(LIBFT_PATH) clean
-# Add any other static library in the same fashion - 6
 	@$(RM) -r $(OBJ_DIR)
 
 fclean: clean
 	@$(MAKE) -C $(LIBFT_PATH) fclean
 	@$(RM) $(NAME)
-	@echo "$(RED)- $(NAME)$(RESET)"
-# Add any other static library in the same fashion - 7
+	@echo Deleting $(NAME)
 
 re: fclean all
