+++ srcs/builtins_2.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   builtins_2.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jorgfern <jorgfern@student.1337.ma>        +#+  +:+       +#+        */
+/*   By: jchamak <jchamak@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/01 13:32:18 by jorgfern          #+#    #+#             */
-/*   Updated: 2023/08/01 13:42:14 by jorgfern         ###   ########.fr       */
+/*   Updated: 2023/08/01 19:08:44 by jchamak          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,7 +61,8 @@ int	ft_builtin_export_2(char *str, char **array)
 {
 	if (ft_strstr(array[1], "="))
 	{
-		if (ft_didnt_expand_exp(array[1]) || array[1][0] == '=' || !ft_isalpha(array[1][0]))
+		if (ft_didnt_expand_exp(array[1]) || array[1][0] == '='
+			|| !ft_isalpha(array[1][0]))
 		{
 			printf("export: `%s': not a valid identifier\n", array[1]);
 			g_all.error = 1;
