+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: jchamak <jchamak@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/01 13:26:10 by jorgfern          #+#    #+#             */
-/*   Updated: 2023/08/01 15:44:02 by jchamak          ###   ########.fr       */
+/*   Updated: 2023/08/01 19:54:55 by jchamak          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -90,8 +90,5 @@ int	main(int argc, char **argv, char **env)
 	ft_var_init();
 	ft_init_env(env);
 	while (1)
-	{
 		ft_readline();
-		system ("leaks minishell");
-	}
 }
